requirements.txt
bitstring
pyserial
